OneDrive - University of Cape Town/Desktop/MobmartAPI/api/apicode/serializers.py
from rest_framework import serializers
from .models import Transactions, User


class TransactionsSerializer(serializers.ModelSerializer):
    class Meta:
        model = Transactions
        fields = "__all__"


class ProfileSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = "__all__"
