OneDrive - University of Cape Town/Desktop/MobmartAPI/api/requirements.txt
asgiref==3.4.1
botocore==1.21.65
cement==2.8.2
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.7
colorama==0.4.3
cryptography==35.0.0
Django==2.2.24
django-pesapal==1.3.1
djangorestframework==3.12.4
future==0.16.0
httplib2==0.20.1
idna==2.8
jmespath==0.10.0
oauth2==1.9.0.post1
pathspec==0.5.9
Pillow==8.4.0
psycopg2==2.9.1
pycparser==2.20
pyOpenSSL==21.0.0
pyparsing==2.4.7
pypiwin32==223
python-dateutil==2.8.2
pytz==2021.3
pywin32==302
PyYAML==5.4.1
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.2
termcolor==1.1.0
urllib3==1.26.7
wcwidth==0.1.9
